chore: 변수 타입 통일
test-order 브랜치와 합병 준비
MenuEntity 수정 및
CategoryEntity 추가

+++ IrumOrder_BackEnd/src/main/java/Irumping/IrumOrder/repository/PayRepository.java
@@ -4,5 +4,5 @@ import Irumping.IrumOrder.entity.OrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PayRepository  extends JpaRepository<OrderEntity, Long> {
-    OrderEntity findByOrderIdAndUserId(int orderId, long userId); // 주문 ID와 사용자 ID로 주문 조회
+    OrderEntity findByOrderIdAndUserId(Integer orderId, Integer userId); // 주문 ID와 사용자 ID로 주문 조회
 }
